Add user input, file checking, additional echo's

+++ macinit.sh
@@ -6,6 +6,21 @@
 # https://github.com/lapwinglabs/blog/blob/master/hacker-guide-to-setting-up-your-mac.md
 
 {
+  # get user options
+  echo "Name: "
+  read -e name
+  echo "Email: "
+  read -e email
+  echo "GitHub username: "
+  read -e username
+  echo "GitHub password: "
+  read -es password
+  echo "ssh key name: "
+  read -e keyname
+  echo "Dotfiles repo path: "
+  read -e dotfiles
+
+
   # install brew
   # check for Homebrew,
   # install if we don't have it
@@ -15,6 +30,7 @@
   fi
 
   # install binaries
+  echo "Installing binaries..."
   binaries=(
     coreutils
     findutils
@@ -35,13 +51,13 @@
     tmux
     vim
   )
-  echo "Installing binaries..."
   brew update
   brew tap homebrew/dupes
   brew install ${binaries[@]}
   brew cleanup
 
   # install Apps
+  echo "Installing apps..."
   apps=(
     dropbox
     evernote
@@ -55,43 +71,71 @@
     transmission
     vlc
   )
-  echo "Installing apps..."
   brew cask install --appdir="/Applications" ${apps[@]}
 
   # setup git
-  git config --global user.name "Brandon Freitag"
-  git config --global user.email "freitagbr@gmail.com"
+  echo "Setting up git..."
+  git config --global user.name "${name}"
+  git config --global user.email "${email}"
 
   # generate ssh key
-  ssh-keygen -t rsa -b 4096 -C "freitagbr@gmail.com"
-
-  # add key to ssh-agent
+  echo "Generating ssh key..."
+  ssh-keygen -t rsa -b 4096 -C "${email}"
   eval "$(ssh-agent -s)"
   ssh-add ~/.ssh/id_rsa
 
   # add key to github
-  curl -u "username" --data '{"title":"${HOSTNAME}","key":"$(cat ~/.ssh/id_rsa.pub)"}' https://api.github.com/user/keys
+  echo "Adding ssh key to GitHub..."
+  curl -u "${username}:${password}" --data '{"title":"${keyname}","key":"$(cat ~/.ssh/id_rsa.pub)"}' https://api.github.com/user/keys
 
   # clone repos
-  git clone https://github.com/freitagbr/dotfiles ~/src/dotfiles
+  echo "Cloning dotfiles repo..."
+  git clone https://github.com/${dotfiles} ~/src/dotfiles
 
   # link the dotfiles
-  ln -s ../../.bash_aliases ~/src/dotfiles/bash_aliases
-  ln -s ../../.bash_colors ~/src/dotfiles/bash_colors
-  ln -s ../../.bash_profile ~/src/dotfiles/bashrc
-  ln -s ../../.gitconfig ~/src/dotfiles/gitconfig
-  ln -s ../../.gitignore ~/src/dotfiles/gitignore
-  ln -s ../../.tmux.conf ~/src/dotfiles/tmux.conf
-  ln -s ../../.vimrc ~/src/dotfiles/vimrc
+  echo "Linking dotfiles..."
+  if [ -f ~/src/dotfiles/bash_aliases ]; then
+    ln -s ../../.bash_aliases ~/src/dotfiles/bash_aliases
+  fi
+  if [ -f ~/src/dotfiles/bash_colors ]; then
+    ln -s ../../.bash_colors ~/src/dotfiles/bash_colors
+  fi
+  if [ -f ~/src/dotfiles/bashrc ]; then
+    ln -s ../../.bash_profile ~/src/dotfiles/bashrc
+  fi
+  if [ -f ~/src/dotfiles/gitconfig ]; then
+    ln -s ../../.gitconfig ~/src/dotfiles/gitconfig
+  fi
+  if [ -f ~/src/dotfiles/gitignore ]; then
+    ln -s ../../.gitignore ~/src/dotfiles/gitignore
+  fi
+  if [ -f ~/src/dotfiles/tmux.conf ]; then
+    ln -s ../../.tmux.conf ~/src/dotfiles/tmux.conf
+  fi
+  if [ -f ~/src/dotfiles/vimrc ]; then
+    ln -s ../../.vimrc ~/src/dotfiles/vimrc
+  fi
 
   # setup vim
+  echo "Setting up vim..."
   git clone https://github.com/VundleVim/Vundle.vim.git ~/.vim/bundle/Vundle.vim
   vim +PluginInstall +qall
 
   # install nvm
+  echo "Installing nvm..."
   curl -o- https://raw.githubusercontent.com/creationix/nvm/v0.31.0/install.sh | bash
   source ~/.bash_profile
   nvm install stable
   nvm use stable
   nvm alias default stable
+
+  # cleanup
+  unset name
+  unset email
+  unset username
+  unset password
+  unset dotfiles
+
+  # done
+  echo "Done."
 }
